Project Report
The objective of this study is to examine COVID-19’s impact on patients with newly discovered liver cancer. This study compares the outcomes from the pre-pandemic period with
those from the pandemic period to ascertain the long-term impact of COVID on patients with
liver cancer. Additionally, it looked at how different patient characteristics and circumstances
linked to these patients’ survival results during the pandemic and pre-pandemic periods.